Changed the GPIO pin type to consume less space. We really aren't going to have greater than 255 pins per port.

+++ Chimera/types.hpp
@@ -122,7 +122,7 @@ namespace Chimera
       Port port          = Port::UNKNOWN_PORT; /**< Pin IO port */
       Drive drive         = Drive::INPUT;       /**< Pin IO drive type */
       State state        = State::LOW;         /**< Default logical pin state on init */
-      uint16_t pin       = 0u;                 /**< Pin number on the given port */
+      uint8_t pin       = 0u;                 /**< Pin number on the given port */
       uint32_t alternate = 0u;                 /**< Project specific indicator of pin AF config*/
     };
